+++ core/dbt/deprecations.py
@@ -62,6 +62,19 @@ class GenerateSchemaNameSingleArgDeprecated(DBTDeprecation):
   '''  # noqa
 
 
+class MaterializationReturnDeprecation(DBTDeprecation):
+    _name = 'materialization-return'
+
+    _description = '''
+    The materialization ("{materialization}") did not explicitly return a list
+    of relations to add to the cache. By default the target relation will be
+    added, but this behavior will be removed in a future version of dbt.
+
+  For more information, see:
+  https://docs.getdbt.com/v0.15/docs/creating-new-materializations#section-6-returning-relations
+    '''.lstrip()
+
+
 _adapter_renamed_description = """\
 The adapter function `adapter.{old_name}` is deprecated and will be removed in
  a future release of dbt. Please use `adapter.{new_name}` instead.
@@ -99,6 +112,7 @@ active_deprecations: Set[str] = set()
 deprecations_list: List[DBTDeprecation] = [
     DBTRepositoriesDeprecation(),
     GenerateSchemaNameSingleArgDeprecated(),
+    MaterializationReturnDeprecation(),
 ]
 
 deprecations: Dict[str, DBTDeprecation] = {