+++ core/dbt/include/global_project/macros/materializations/snapshot/snapshot.sql
@@ -266,4 +266,6 @@
 
   {{ run_hooks(post_hooks, inside_transaction=False) }}
 
+  {{ return({'relations': [target_relation]}) }}
+
 {% endmaterialization %}