+++ core/dbt/include/global_project/macros/materializations/view/create_or_replace_view.sql
@@ -58,4 +58,7 @@
       -- No transactions on BigQuery
       {{ run_hooks(post_hooks, inside_transaction=False) }}
   {% endif %}
+
+  {{ return({'relations': [target_relation]}) }}
+
 {% endmacro %}